refactor: cleanup headers
